+++ controllers/signup.go
@@ -6,16 +6,10 @@ import (
 	"net/http"
 )
 
-type RegisterInput struct {
-	Name     string `json:"name" binding:"required"`
-	Email    string `json:"email" binding:"required"`
-	Password string `json:"password" binding:"required"`
-}
-
-// register function will register the detail of user
+// Register function will register the detail of user
 func Register(c *gin.Context) {
 
-	var input RegisterInput
+	var input models.RegisterInput
 
 	if err := c.ShouldBindJSON(&input); err != nil {
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
